Move rvmg function to custom folder
I had originally added the rvmg function directly in plugins/rvm/rvm.plugin.zsh, but when oh-my-zsh tried to auto update, it prevented the update because I hadn't commited that change and git won't allow a merge in that case.

Committing the change will allow the update, but the right thing to do is put any custom functions and aliases in the custom directory.